+++ internal/model/user.go
@@ -15,7 +15,7 @@ type (
 	}
 	UserBalance struct {
 		Current   float32 `json:"current"`
-		Withdrawn int     `json:"withdrawn"`
+		Withdrawn float32 `json:"withdrawn"`
 	}
 )
